Use CDAF managed windows image

+++ automation/provisioning/newForest.ps1
@@ -9,6 +9,34 @@ function executeExpression ($expression) {
     if ( $error[0] ) { Write-Host "[$scriptName] `$error[0] = $error"; exit 3 }
 }
 
+function executeRetry ($expression) {
+	$exitCode = 1
+	$wait = 10
+	$retryMax = 10
+	$retryCount = 0
+	while (( $retryCount -le $retryMax ) -and ($exitCode -ne 0)) {
+		$exitCode = 0
+		$error.clear()
+		Write-Host "[$scriptName][$retryCount] $expression"
+		try {
+			Invoke-Expression $expression
+		    if(!$?) { Write-Host "[$scriptName] `$? = $?"; $exitCode = 1 }
+		} catch { echo $_.Exception|format-list -force; $exitCode = 2 }
+	    if ( $error[0] ) { Write-Host "[$scriptName] `$error[0] = $error"; $exitCode = 3 }
+		if ( $LASTEXITCODE -eq 3010 ) { $LASTEXITCODE = 0 } # 3010 is a normal exit
+	    if ( $lastExitCode -ne 0 ) { Write-Host "[$scriptName] `$lastExitCode = $lastExitCode "; $exitCode = $lastExitCode }
+	    if ($exitCode -ne 0) {
+			if ($retryCount -ge $retryMax ) {
+				Write-Host "[$scriptName] Retry maximum ($retryCount) reached, exiting with code $exitCode"; exit $exitCode
+			} else {
+				$retryCount += 1
+				Write-Host "[$scriptName] Wait $wait seconds, then retry $retryCount of $retryMax"
+				sleep $wait
+			}
+		}
+    }
+}
+
 # Create or reuse mount directory
 function mountWim ($media, $wimIndex, $mountDir) {
 	Write-Host "[$scriptName] Validate WIM source ${media}:${wimIndex} using Deployment Image Servicing and Management (DISM)"
@@ -62,8 +90,14 @@ if ($controlReboot) {
 	$controlReboot = 'yes'
     Write-Host "[$scriptName] controlReboot : $controlReboot (default)"
 }
-if ($controlReboot -eq 'yes') {
+if ($controlReboot -eq 'no') {
+    Write-Host "`n[$scriptName] controlReboot is $controlReboot, allow Active Directory feature to reboot"
+} else {
 	$rebootOption = '-NoRebootOnCompletion'
+}
+# Provisionig Script builder
+if ( $env:PROV_SCRIPT_PATH ) {
+	Add-Content "$env:PROV_SCRIPT_PATH" "executeExpression `"./automation/provisioning/$scriptName $forest ********** $media $wimIndex $controlReboot`""
 }
 
 # Cannot run interactive via remote PowerShell
@@ -120,19 +154,27 @@ if ( Test-Path $media ) {
 Write-Host
 Write-Host "[$scriptName] Install Active Directory Domain Roles and Services using Deployment Image Servicing and Management (DISM)"
 Write-Host
-Write-Host "[$scriptName] Source not required for RSAT (Remote Server Administration Tools)"
+Write-Host "[$scriptName]   Remote Server Administration Tools (RSAT)"
 Write-Host
 $featureList = @('ServerManager-Core-RSAT', 'ServerManager-Core-RSAT-Role-Tools', 'RSAT-AD-Tools-Feature')
 foreach ($feature in $featureList) {
-	executeExpression "dism /online /NoRestart /enable-feature /featurename:$feature /LimitAccess /Quiet"
+	executeExpression "dism /online /NoRestart /enable-feature /featurename:$feature $sourceOption"
+	if ( $lastExitCode -ne 0 ) {
+		Write-Host "[$scriptName] DISM failed with `$lastExitCode = $lastExitCode, retry from WSUS/Internet"
+		executeRetry "dism /online /NoRestart /enable-feature /featurename:$feature /Quiet"
+	}
 }
 
 Write-Host
-Write-Host "[$scriptName] Source required for Directory Services"
+Write-Host "[$scriptName]   Source required for Directory Services"
 Write-Host
 $featureList = @('ActiveDirectory-PowerShell', 'DirectoryServices-DomainController', 'RSAT-ADDS-Tools-Feature', 'DirectoryServices-DomainController-Tools', 'DNS-Server-Full-Role', 'DNS-Server-Tools', 'DirectoryServices-AdministrativeCenter')
 foreach ($feature in $featureList) {
 	executeExpression "dism /online /NoRestart /enable-feature /featurename:$feature $sourceOption"
+	if ( $lastExitCode -ne 0 ) {
+		Write-Host "[$scriptName] DISM failed with `$lastExitCode = $lastExitCode, retry from WSUS/Internet"
+		executeRetry "dism /online /NoRestart /enable-feature /featurename:$feature /Quiet"
+	}
 }
 
 if ( Test-Path "$defaultMount\windows" ) {
@@ -143,24 +185,26 @@ if ( Test-Path "$defaultMount\windows" ) {
 # https://github.com/rgl/windows-domain-controller-vagrant/blob/master/provision/domain-controller.ps1
 #   If using -NoRebootOnCompletion do not use reload module in Vagrant or it will fail (raise_if_auth_error)
 
-Write-Host
-Write-Host "[$scriptName] Create the new Forest and convert this host into the FSMO Domain Controller"
-Write-Host
-$securePassword = ConvertTo-SecureString $password -asplaintext -force
-executeExpression "Install-ADDSForest -Force $rebootOption -DomainName `"$forest`" -SafeModeAdministratorPassword `$securePassword"
+if ( $controlReboot -eq 'none' ) {
+	Write-Host "`n[$scriptName] controlReboot is $controlReboot, skipping forest configuration.`n"
+} else {
+	Write-Host "`n[$scriptName] Create the new Forest and convert this host into the FSMO Domain Controller`n"
+	$securePassword = ConvertTo-SecureString $password -asplaintext -force
+	executeExpression "Install-ADDSForest -Force $rebootOption -DomainName `"$forest`" -SafeModeAdministratorPassword `$securePassword"
 
-# https://github.com/dbroeglin/windows-lab/blob/master/provision/02_install_forest.ps1
-#   Tried putting a sleep in, but reboot still triggered a Vagrant error
-#   rescue in block in parse_header': HTTPClient::KeepAliveDisconnected: An existing connection was forcibly closed by the remote host. @ io_fillbuf - fd:3  (HTTPClient::KeepAliveDisconnected)
-#   Write-Host "Start sleeping until reboot to prevent vagrant connection failures..."
-#   executeExpression "Start-Sleep 180"
+	# https://github.com/dbroeglin/windows-lab/blob/master/provision/02_install_forest.ps1
+	#   Tried putting a sleep in, but reboot still triggered a Vagrant error
+	#   rescue in block in parse_header': HTTPClient::KeepAliveDisconnected: An existing connection was forcibly closed by the remote host. @ io_fillbuf - fd:3  (HTTPClient::KeepAliveDisconnected)
+	#   Write-Host "Start sleeping until reboot to prevent vagrant connection failures..."
+	#   executeExpression "Start-Sleep 180"
 
-# https://groups.google.com/forum/#!topic/vagrant-up/JNMOCYpHSt8
-#   This attempt using chef_solo also appears to fail with the same problem.
+	# https://groups.google.com/forum/#!topic/vagrant-up/JNMOCYpHSt8
+	#   This attempt using chef_solo also appears to fail with the same problem.
 
-if ($controlReboot -eq 'yes') {
-	Write-Host "[$scriptName] Controlled Reboot requested, initiating reboot ..."
-	executeExpression "shutdown /r /t 0"
+	if ($controlReboot -eq 'yes') {
+		Write-Host "[$scriptName] Controlled Reboot requested, initiating reboot ..."
+		executeExpression "shutdown /r /t 0"
+	}
 }
 
 Write-Host
